Refactor Qwen3 embedding by moving pooling functions to a new module
